+++ envcoord/mailexpander/expander.py
@@ -12,12 +12,10 @@ import getopt
 import ldap
 import logging
 import os
-import re
 import smtplib
 import string
 import sys
 import time
-from lxml import html
 
 __version__ = """$Id: expander.py 40888 2017-04-05 09:47:09Z tiberich $"""
 
@@ -196,6 +194,18 @@ class Expander(object):
                         return RETURN_CODES['EX_CONFIG']
             return RETURN_CODES['EX_OK']
 
+        # Check if role is deactivated
+        if self.is_deactivated(role_data):
+            log.info("Role %s is deactivated, rejecting email from %s",
+                     role, from_email)
+            if not debug_mode:
+                try:
+                    self.send_deactivated_email(from_email, role, role_email)
+                except Exception:
+                    log.exception("Error sending deactivation notice")
+                    return RETURN_CODES['EX_TEMPFAIL']
+            return RETURN_CODES['EX_OK']
+
         # Check if from_email can expand
         if self.can_expand(from_email, role, role_data) is False:
             log.info("%s cannot send to %s@envcoord", from_email, role)
@@ -290,6 +300,10 @@ class Expander(object):
             if self.also_send_to != ['']:
                 retval = self.send_emails(
                     role_email, self.also_send_to, content)
+                if retval != RETURN_CODES['EX_OK']:
+                    log.error("Error %s while sending to %s",
+                              retval, self.also_send_to)
+                    return retval
 
             # Send e-mails
             for emails in email_batches:
@@ -297,7 +311,7 @@ class Expander(object):
                     role_email, emails, content)
                 if retval != RETURN_CODES['EX_OK']:
                     log.error("Error %s while sending to %s",
-                              retval, self.no_owner_send_to)
+                              retval, emails)
                     return retval
             if not self.skip_confirmation_email:
                 try:
@@ -318,8 +332,8 @@ class Expander(object):
         # this allows "inheriting" permissions from above roles
         # see http://taskman.eionet.europa.eu/issues/20422
 
-        if 'permittedperson' not in role_data:
-            role_data['permittedperson'] = []
+        if 'permittedPerson' not in role_data:
+            role_data['permittedPerson'] = []
         if 'permittedSender' not in role_data:
             role_data['permittedSender'] = []
 
@@ -358,13 +372,13 @@ class Expander(object):
 
             for person_dn in role_info.get('permittedPerson', []):
                 try:
-                    email = self.agent._query(person_dn)['mail'][0]
+                    sender_email = self.agent._query(person_dn)['mail'][0]
                 except Exception as e:
                     # Log that we couldn't get the email.
-                    log.exception("Invalid DN: %s; %s" (person_dn, e))
+                    log.exception("Invalid DN: %s; %s" % (person_dn, e))
                     continue
                 else:
-                    senders.add(email)
+                    senders.add(sender_email)
 
         role_data['permittedSender'] = filter(None, set(senders))
 
@@ -421,7 +435,7 @@ class Expander(object):
                             except Exception as e:
                                 # Log that we couldn't get the email.
                                 log.exception(
-                                    "Invalid `owner` DN: %s; %s" (owner_dn, e))
+                                    "Invalid `owner` DN: %s; %s" % (owner_dn, e))
                                 continue
                             if from_email in map(str.lower, owner['mail']):
                                 return True
@@ -447,6 +461,44 @@ class Expander(object):
                     continue
 
         return False
+
+    def is_deactivated(self, role_data):
+        """ Check if a role is deactivated based on the LDAP 'l' attribute.
+        The attribute stores values like 'deactivated:True' or
+        'deactivated:False'.
+        """
+        raw = role_data.get('l', ['deactivated:False'])[0]
+        if isinstance(raw, bytes):
+            raw = raw.decode('utf-8')
+        return raw.lower().split('deactivated:')[-1] == 'true'
+
+    def send_deactivated_email(self, to_email, role, role_email):
+        """ Send a notification to the sender that the role is deactivated
+        and no email will be delivered.
+        """
+        log.info("Sending deactivation notice to %s for role %s",
+                 to_email, role)
+        content = ""
+        template_path = os.path.join(os.path.dirname(__file__),
+                                     'templates',
+                                     'deactivated_role_email.html')
+        if os.path.isfile(template_path):
+            f = open(template_path, 'rb')
+            content = f.read()
+            f.close()
+
+        content = content.replace('{{role_id}}', role)
+        content = content.replace('{{role_email}}', role_email)
+
+        html_part = MIMEText(content, 'html')
+        message = MIMEMultipart('alternative')
+        message['Subject'] = ("Email not delivered: role %s is deactivated"
+                              % role)
+        message['From'] = self.noreply
+        message['To'] = to_email
+        message.attach(html_part)
+
+        return self.send_emails(self.noreply, [to_email], message.as_string())
 
     def send_confirmation_email(self, subject, to_email, role):
         """ If sending emails succeeded send a confirmation email to sender and